docs: move theme.ts to constants dir
